staticdatagen::generators
Module news_sitemap

The news_sitemap module contains the news sitemap generator. News Sitemap Generation Module

This module handles the creation and generation of Google News sitemaps, which help search engines discover and index news content. It follows the Google News Sitemap protocol specification.

Features

Creation of news sitemap data structures from metadata

Validation of news publication dates

Proper XML formatting for news sitemaps

Support for news-specific metadata (genres, keywords, etc.)

Example

use std::collections::HashMap;
use staticdatagen::generators::news_sitemap::{NewsSiteMapConfig, NewsSiteMapGenerator};

let mut metadata = HashMap::new();
metadata.insert("news_title".to_string(), "Breaking News".to_string());
metadata.insert(
    "news_publication_date".to_string(),
    "Tue, 20 Feb 2024 15:15:15 GMT".to_string(),
);

let config = NewsSiteMapConfig::new(metadata);
let generator = NewsSiteMapGenerator::new(config);
let news_sitemap = generator.generate_xml();

Structs

NewsSiteMapConfig

Configuration for generating a news sitemap.

NewsSiteMapGenerator

Generator for creating a news sitemap.