staticdatagen::modules::news_sitemap
Function create_news_site_map_data

pub fn create_news_site_map_data(metadata: &HashMap<String, String>) -> NewsData

Creates a NewsData object from metadata.

This function processes metadata to create a news sitemap entry. All dates are converted to the required format and content is validated.

Arguments

metadata - A reference to a HashMap containing news metadata

Returns

NewsData - A struct containing the news sitemap information

Example

use std::collections::HashMap;
use staticdatagen::modules::news_sitemap::create_news_site_map_data;

let mut metadata = HashMap::new();
metadata.insert("news_title".to_string(), "Breaking News".to_string());

let news_data = create_news_site_map_data(&metadata);