staticdatagen::modules::preprocessor
Function preprocess_content

pub fn preprocess_content(
    content: &str,
    class_regex: &Regex,
    img_regex: &Regex,
) -> Result<String, Box<dyn Error>>

Preprocesses the Markdown content to update class attributes and image tags.

Arguments

content - A string containing the Markdown content to be processed.

class_regex - A reference to a Regex object for matching class attributes.

img_regex - A reference to a Regex object for matching image tags.

Returns

A Result containing a String with the processed Markdown content, or a Box<dyn Error> if an error occurs.