staticdatagen::modules::tags
Function generate_tags

pub fn generate_tags(
    file: &FileData,
    metadata: &HashMap<String, String>,
) -> HashMap<String, Vec<HashMap<String, String>>>

Generates a tag list from the given FileData and metadata, and returns it as a HashMap.

Arguments

file - A reference to a FileData struct, which contains the content of a single file.

metadata - A reference to a HashMap containing metadata like tags, title, etc.

Returns

A HashMap mapping each tag to a vector of HashMaps containing associated data like title, description, etc.

Examples

use staticdatagen::models::data::FileData;
use staticdatagen::modules::tags::generate_tags;
use std::collections::HashMap;

let file = FileData { content: "This is a test".to_string(), ..Default::default() };
let mut metadata = HashMap::new();
metadata.insert("tags".to_string(), "tag1,tag2".to_string());

let result = generate_tags(&file, &metadata);