staticdatagen::modules::tags
Function generate_tags_html

pub fn generate_tags_html(
    global_tags_data: &HashMap<String, Vec<PageData>>,
) -> String

Generates the HTML content for displaying tags and their associated pages.

This function takes a HashMap that maps tags to a list of PageData objects, where each PageData object contains metadata about a page that uses the tag. It then generates HTML content that displays these tags along with the pages that use them.

Arguments

global_tags_data - A reference to a HashMap where each key is a tag and the corresponding value is a Vec of PageData objects that use the tag.

Returns

A String containing the generated HTML content.

Examples

use std::collections::HashMap;
use staticdatagen::models::data::PageData;
use staticdatagen::modules::tags::generate_tags_html;

let mut global_tags_data = HashMap::new();
global_tags_data.insert(
    "tag1".to_string(),
    vec![
        PageData {
            date: "2022-09-01".to_string(),
            description: "Description 1".to_string(),
            permalink: "/page1".to_string(),
            title: "Page 1".to_string(),
        },
    ],
);

global_tags_data.insert(
    "tag2".to_string(),
    vec![
        PageData {
            date: "2022-09-02".to_string(),
            description: "Description 2".to_string(),
            permalink: "/page2".to_string(),
            title: "Page 2".to_string(),
        },
    ],
);

let html_content = generate_tags_html(&global_tags_data);