staticdatagen::utilities::backup
Function backup_file

pub fn backup_file(file_path: &Path) -> Result<PathBuf>

Creates a backup of a file by copying it with a “.src.html” extension.

Arguments

file_path - A reference to the path of the file to backup

Returns

io::Result<PathBuf> - The path to the created backup file

Examples

use std::path::Path;
use staticdatagen::utilities::backup::backup_file;
let file_path = Path::new("myfile.txt");
match backup_file(file_path) {
    Ok(backup_path) => println!("Backup created at: {:?}", backup_path),
    Err(e) => eprintln!("Failed to create backup: {}", e),
}

Errors

Returns an error if:

The source file doesn’t exist

There are insufficient permissions

The backup operation fails due to I/O issues