staticdatagen::utilities::directory
Function truncate

pub fn truncate(path: &Path, length: usize) -> Option<String>

Truncates a path to only have a set number of path components.

Arguments

path - The path to truncate.

length - The number of path components to keep.

Returns

An Option<String> containing the truncated path, or None if not truncated.

Examples

use std::path::Path;
use staticdatagen::utilities::directory::truncate;

let path = Path::new("/a/b/c/d/e");
assert_eq!(truncate(path, 3), Some("c/d/e".to_string()));