staticdatagen::utilities::element
Function write_element

pub fn write_element(
    writer: &mut Writer<Cursor<Vec<u8>>>,
    name: &str,
    value: &str,
) -> Result<(), Box<dyn Error>>

Helper function to write XML element

This function takes a reference to a Writer object, a string containing the name of the element, and a string containing the value of the element,

Arguments

writer - A reference to a Writer object.

name - A string containing the name of the element.

value - A string containing the value of the element.

Returns

Result<(), Box<dyn std::error::Error>> - A result indicating success or failure.

Ok(()) if the element was written successfully.

Err(Box<dyn std::error::Error>) if an error occurred during the writing process.